requirements.txt
pdf2image
img2pdf
Pillow
numpy
customtkinter
